api/flask/requirements.txt
beautifulsoup4==4.12.2
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cryptography==41.0.3
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==23.2.1
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.29.0
setuptools==68.0.0
six==1.16.0
soupsieve==2.4
urllib3==1.26.16
uWSGI==2.0.21
Werkzeug==2.2.3
wheel==0.41.2
